feat: alterando tipo de dado financeiro de Double para BidDecimal

+++ src/main/java/br/com/dbccompany/sistemadereembolso/Sistema/de/reembolso/dto/reembolso/ReembolsoDTO.java
@@ -4,6 +4,8 @@ import br.com.dbccompany.sistemadereembolso.Sistema.de.reembolso.dto.arquivos.An
 import br.com.dbccompany.sistemadereembolso.Sistema.de.reembolso.dto.usuario.UsuarioComposeDTO;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -18,7 +20,7 @@ public class ReembolsoDTO {
     private LocalDateTime dataUltimaAlteracao;
 
     @Schema(description = "Valor do reembolso")
-    private Double valor;
+    private BigDecimal valor;
 
     @Schema(description = "Status do reembolso")
     private String statusDoReembolso;
